+++ src/main/java/com/cluting/clutingbackend/plan/domain/enums/Role.java
@@ -1,0 +1,5 @@
+package com.cluting.clutingbackend.plan.domain.enums;
+
+public enum Role {
+    GENERAL, ADMIN
+}
