--- src/main/java/com/cluting/clutingbackend/user/domain/enums/Role.java
@@ -1,5 +1,0 @@
-package com.cluting.clutingbackend.user.domain.enums;
-
-public enum Role {
-    GENERAL, ADMIN
-}
